add 10 other classes for style check
