+++ iotdb/src/main/java/org/apache/iotdb/db/query/externalsort/serialize/TimeValuePairDeserializer.java
@@ -1,34 +1,31 @@
 /**
  * Copyright © 2019 Apache IoTDB(incubating) (dev@iotdb.apache.org)
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  */
+
 package org.apache.iotdb.db.query.externalsort.serialize;
 
-import org.apache.iotdb.db.utils.TimeValuePair;
-
 import java.io.IOException;
+import org.apache.iotdb.db.utils.TimeValuePair;
 
 public interface TimeValuePairDeserializer {
 
-    boolean hasNext() throws IOException;
+  boolean hasNext() throws IOException;
 
-    TimeValuePair next() throws IOException;
+  TimeValuePair next() throws IOException;
 
-    /**
-     * Close current deserializer
-     * 
-     * @throws IOException
-     */
-    void close() throws IOException;
+  /**
+   * Close current deserializer.
+   *
+   */
+  void close() throws IOException;
 }
