+++ iotdb/src/main/java/org/apache/iotdb/db/query/externalsort/serialize/TimeValuePairSerializer.java
@@ -1,26 +1,25 @@
 /**
  * Copyright © 2019 Apache IoTDB(incubating) (dev@iotdb.apache.org)
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  */
+
 package org.apache.iotdb.db.query.externalsort.serialize;
 
-import org.apache.iotdb.db.utils.TimeValuePair;
 import java.io.IOException;
+import org.apache.iotdb.db.utils.TimeValuePair;
 
 public interface TimeValuePairSerializer {
 
-    void write(TimeValuePair timeValuePair) throws IOException;
+  void write(TimeValuePair timeValuePair) throws IOException;
 
-    void close() throws IOException;
+  void close() throws IOException;
 }
