+++ tsfile/src/main/java/org/apache/iotdb/tsfile/read/query/timegenerator/TimeGenerator.java
@@ -1,37 +1,34 @@
 /**
  * Copyright © 2019 Apache IoTDB(incubating) (dev@iotdb.apache.org)
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  */
+
 package org.apache.iotdb.tsfile.read.query.timegenerator;
 
-import org.apache.iotdb.tsfile.read.common.Path;
-import org.apache.iotdb.tsfile.read.common.Path;
-
 import java.io.IOException;
+import org.apache.iotdb.tsfile.read.common.Path;
 
 /**
- * All SingleSeriesExpression involved in a IExpression will be transferred to a TimeGenerator tree whose leaf nodes are
- * all SeriesReaders, The TimeGenerator tree can generate the next timestamp that satisfies the filter condition.
- *
+ * All SingleSeriesExpression involved in a IExpression will be transferred to a TimeGenerator tree
+ * whose leaf nodes are all SeriesReaders, The TimeGenerator tree can generate the next timestamp
+ * that satisfies the filter condition.
  * Then we use this timestamp to get values in other series that are not included in IExpression
  */
 public interface TimeGenerator {
 
-    boolean hasNext() throws IOException;
+  boolean hasNext() throws IOException;
 
-    long next() throws IOException;
+  long next() throws IOException;
 
-    Object getValue(Path path, long time) throws IOException;
+  Object getValue(Path path, long time) throws IOException;
 
 }
